final COMMIT

+++ src/main/java/com/keyin/domain/golfmember/MemberService.java
@@ -1,5 +1,6 @@
 package com.keyin.domain.golfmember;
 
+import com.keyin.domain.golftournament.Tournament;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.time.LocalDate;
@@ -19,7 +20,11 @@ import java.util.List;
         return (List<GolfMember>) memberRepository.saveAll(members);
     }
 
-    public List <GolfMember> searchByName(String name) {
+    public List<GolfMember> findAllGolfMembers() {
+        return (List<GolfMember>) memberRepository.findAll();
+    }
+
+    public List<GolfMember> searchByName(String name) {
         return memberRepository.findByNameContaining(name);
     }
 
@@ -36,21 +41,42 @@ import java.util.List;
     }
 
     public GolfMember updateGolfMember(GolfMember updatedGolfMember) {
-        GolfMember golfMemberToUpdate = findById(updatedGolfMember.getId());
+        GolfMember existingMember = findById(updatedGolfMember.getId());
 
-        if (golfMemberToUpdate != null) {
-            golfMemberToUpdate.setName(updatedGolfMember.getName());
-            golfMemberToUpdate.setAddress(updatedGolfMember.getAddress());
-            golfMemberToUpdate.setEmail(updatedGolfMember.getEmail());
-            golfMemberToUpdate.setPhone(updatedGolfMember.getPhone());
-            golfMemberToUpdate.setStartDate(updatedGolfMember.getStartDate());
-            golfMemberToUpdate.setMembershipDuration(updatedGolfMember.getMembershipDuration());
-            return memberRepository.save(golfMemberToUpdate);
+        if (existingMember != null) {
+            if (updatedGolfMember.getAddress() != null) {
+                existingMember.setAddress(updatedGolfMember.getAddress());
+            }
+            if (updatedGolfMember.getName() != null) {
+                existingMember.setName(updatedGolfMember.getName());
+            }
+            if (updatedGolfMember.getEmail() != null) {
+                existingMember.setEmail(updatedGolfMember.getEmail());
+            }
+            if (updatedGolfMember.getPhone() != null) {
+                existingMember.setPhone(updatedGolfMember.getPhone());
+            }
+            if (updatedGolfMember.getStartDate() != null) {
+                existingMember.setStartDate(updatedGolfMember.getStartDate());
+            }
+            if (updatedGolfMember.getMembershipDuration() != null) {
+                existingMember.setMembershipDuration(updatedGolfMember.getMembershipDuration());
+            }
+            return memberRepository.save(existingMember);
         }
         return null;
     }
-    public void deleteMember(Long id) {
-        memberRepository.deleteById(id);
+
+    public void deleteGolfMember(Long id) {
+        GolfMember member = findById(id);
+        if (member != null) {
+            // Remove the member from all associated tournaments
+            for (Tournament tournament : member.getTournaments()) {
+                tournament.getParticipatingMembers().remove(member);
+            }
+            member.getTournaments().clear();
+            memberRepository.deleteById(id);
+        }
     }
 }
